feat: ✨ Extract libmemory
