+++ Cargo.toml
@@ -11,6 +11,7 @@ members = [
   "libs/local-irq",
   "libs/locking",
   "libs/log",
+  "libs/memory",
   "libs/primitives",
   "libs/qemu",
   "libs/test",
@@ -45,6 +46,7 @@ libkernel-state = { path = "libs/kernel-state" }
 liblocal-irq = { path = "libs/local-irq" }
 liblocking = { path = "libs/locking" }
 liblog = { path = "libs/log" }
+libmemory = { path = "libs/memory" }
 libprimitives = { path = "libs/primitives" }
 libqemu = { path = "libs/qemu" }
 libtest = { path = "libs/test" }
