+++ libs/memory/Cargo.toml
@@ -1,0 +1,39 @@
+[package]
+name = "libmemory"
+
+description = "Vesper nanokernel Memory Descriptors shared code library, useful also for the ttt tool."
+
+authors = { workspace = true }
+categories = { workspace = true }
+documentation = { workspace = true }
+edition = { workspace = true }
+homepage = { workspace = true }
+license = { workspace = true }
+readme = { workspace = true }
+repository = { workspace = true }
+version = { workspace = true }
+
+publish = false
+
+[badges]
+maintenance = { status = "experimental" }
+
+[dependencies]
+aarch64-cpu = { workspace = true }
+bit_field = { workspace = true }
+bitflags = { workspace = true }
+buddy-alloc = { workspace = true }
+cfg-if = { workspace = true }
+liblocking = { workspace = true }
+liblog = { workspace = true }
+once_cell = { workspace = true }
+snafu = { workspace = true }
+tock-registers = { workspace = true }
+usize_conversions = { workspace = true }
+ux = { workspace = true }
+
+[lib]
+test = false
+
+[lints]
+workspace = true
