+++ libs/memory/src/arch/aarch64/addr/mod.rs
@@ -3,6 +3,5 @@
  * Copyright (c) Berkus Decker <berkus+vesper@metta.systems>
  */
 
-mod asid;
-
-pub use asid::*;
+pub mod asid;
+// pub use asid::*;
