+++ libs/memory/src/arch/mod.rs
@@ -1,0 +1,5 @@
+#[cfg(target_arch = "aarch64")]
+pub mod aarch64;
+
+#[cfg(target_arch = "aarch64")]
+pub use aarch64::*;
