+++ libs/memory/src/mm/bump_allocator.rs
@@ -4,6 +4,7 @@
  */
 
 // @todo Use alloc-fmt crate for logging in allocators
+// @todo move to liballoc or sth, outside of the kernel
 
 use {
     core::{
@@ -27,8 +28,8 @@ unsafe impl Allocator for BumpAllocator {
         let end = start + layout.size();
 
         println!(
-            "[i] {}:\n    Allocating Start {:#010x} End {:#010x}",
-            self.name, start, end
+            "[i] {}:\n    Allocating Start {start:#010x} End {end:#010x}",
+            self.name,
         );
 
         if end > self.pool_end {
@@ -62,32 +63,5 @@ impl BumpAllocator {
             pool_end,
             name,
         }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    // Validate allocator allocates from the provided address range
-    // Validate allocation fails when range is exhausted
-    #[test_case]
-    fn test_allocates_within_init_range() {
-        let allocator = BumpAllocator::new(256, 512, "Test allocator 1");
-        let result1 = allocator.allocate(unsafe { Layout::from_size_align_unchecked(128, 1) });
-        assert!(result1.is_ok());
-        let result2 = allocator.allocate(unsafe { Layout::from_size_align_unchecked(128, 32) });
-        println!("{:?}", result2);
-        assert!(result2.is_ok());
-        let result3 = allocator.allocate(unsafe { Layout::from_size_align_unchecked(1, 1) });
-        assert!(result3.is_err());
-    }
-    // Creating with end <= start sshould fail
-    // @todo return Result<> from new?
-    #[test_case]
-    fn test_bad_allocator() {
-        let bad_allocator = BumpAllocator::new(512, 256, "Test allocator 2");
-        let result1 = bad_allocator.allocate(unsafe { Layout::from_size_align_unchecked(1, 1) });
-        assert!(result1.is_err());
     }
 }
