+++ libs/memory/src/mmu/translation_table.rs
@@ -1,11 +1,11 @@
 //! Translation table.
 
 #[cfg(target_arch = "aarch64")]
-use crate::arch::aarch64::memory::mmu::translation_table as arch_translation_table;
+use crate::arch::aarch64::mmu::translation_table as arch_translation_table;
 
 use {
     super::{AttributeFields, MemoryRegion},
-    crate::memory::{Address, Physical, Virtual},
+    crate::{Address, Physical, Virtual},
 };
 
 //--------------------------------------------------------------------------------------------------
@@ -50,47 +50,5 @@ pub mod interface {
             phys_region: &MemoryRegion<Physical>,
             attr: &AttributeFields,
         ) -> Result<(), &'static str>;
-    }
-}
-
-//--------------------------------------------------------------------------------------------------
-// Testing
-//--------------------------------------------------------------------------------------------------
-
-#[cfg(test)]
-mod tests {
-    use {
-        super::*,
-        crate::memory::mmu::{AccessPermissions, MemAttributes, PageAddress},
-        arch_translation_table::MinSizeTranslationTable,
-        interface::TranslationTable,
-    };
-
-    /// Sanity checks for the TranslationTable implementation.
-    #[test_case]
-    fn translation_table_implementation_sanity() {
-        // This will occupy a lot of space on the stack.
-        let mut tables = MinSizeTranslationTable::new();
-
-        tables.init();
-
-        let virt_start_page_addr: PageAddress<Virtual> = PageAddress::from(0);
-        let virt_end_exclusive_page_addr: PageAddress<Virtual> =
-            virt_start_page_addr.checked_offset(5).unwrap();
-
-        let phys_start_page_addr: PageAddress<Physical> = PageAddress::from(0);
-        let phys_end_exclusive_page_addr: PageAddress<Physical> =
-            phys_start_page_addr.checked_offset(5).unwrap();
-
-        let virt_region = MemoryRegion::new(virt_start_page_addr, virt_end_exclusive_page_addr);
-        let phys_region = MemoryRegion::new(phys_start_page_addr, phys_end_exclusive_page_addr);
-
-        let attr = AttributeFields {
-            mem_attributes: MemAttributes::CacheableDRAM,
-            acc_perms: AccessPermissions::ReadWrite,
-            execute_never: true,
-        };
-
-        unsafe { assert_eq!(tables.map_at(&virt_region, &phys_region, &attr), Ok(())) };
     }
 }
