+++ libs/memory/src/platform/README.md
@@ -1,0 +1,1 @@
+Temp, until libplatform is introduced.
