+++ libs/memory/src/platform/mod.rs
@@ -1,0 +1,4 @@
+pub type KernelGranule = crate::mmu::TranslationGranule<{ 64 * 1024 }>;
+
+pub mod raspberrypi;
+pub use raspberrypi::*;
