+++ machine/src/arch/aarch64/mod.rs
@@ -6,4 +6,4 @@
 //! Implementation of aarch64 kernel functions.
 
 pub mod cpu;
-pub mod memory;
+// pub mod traps;
